+++ src/main/java/dev/coder2195/rewired/registry/RewiredBlocks.java
@@ -7,18 +7,18 @@ import net.minecraft.references.BlockItemId;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 //? fabric {
-/*import net.minecraft.core.Registry;
+import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
-*///? } else {
-import static dev.coder2195.rewired.Rewired.MOD_ID;
+//? } else {
+/*import static dev.coder2195.rewired.Rewired.MOD_ID;
 import net.neoforged.neoforge.registries.DeferredRegister;
-//? }
+*///? }
 
 import java.util.function.Function;
 
 public interface RewiredBlocks {
 	//? neoforge
-	DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(Rewired.MOD_ID);
+	//DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(Rewired.MOD_ID);
 
 	BlockItemId AND_GATE_ID = blockItem("and_gate");
 	BlockItemId OR_GATE_ID = blockItem("or_gate");
@@ -27,6 +27,8 @@ public interface RewiredBlocks {
 	BlockItemId NOR_GATE_ID = blockItem("nor_gate");
 	BlockItemId XNOR_GATE_ID = blockItem("xnor_gate");
 	BlockItemId AVERAGE_GATE_ID = blockItem("average_gate");
+	BlockItemId MIN_GATE_ID = blockItem("min_gate");
+	BlockItemId MAX_GATE_ID = blockItem("max_gate");
 	static BlockItemId blockItem(String id) {
 		return BlockItemId.create(Rewired.id(id), Rewired.id(id));
 	}
@@ -38,6 +40,8 @@ public interface RewiredBlocks {
 	Holder<Block> NOR_GATE = register(NOR_GATE_ID, NorGateBlock::new, BlockBehaviour.Properties.of());
 	Holder<Block> XNOR_GATE = register(XNOR_GATE_ID, XnorGateBlock::new, BlockBehaviour.Properties.of());
 	Holder<Block> AVERAGE_GATE = register(AVERAGE_GATE_ID, AverageGateBlock::new, BlockBehaviour.Properties.of());
+	Holder<Block> MIN_GATE = register(MIN_GATE_ID, MinGateBlock::new, BlockBehaviour.Properties.of());
+	Holder<Block> MAX_GATE = register(MAX_GATE_ID, MaxGateBlock::new, BlockBehaviour.Properties.of());
 
 	Holder<Block>[] GATES = new Holder[]{
 		AND_GATE,
@@ -46,7 +50,9 @@ public interface RewiredBlocks {
 		NAND_GATE,
 		NOR_GATE,
 		XNOR_GATE,
-		AVERAGE_GATE
+		AVERAGE_GATE,
+		MIN_GATE,
+		MAX_GATE
 	};
 
 	static Holder<Block> register(BlockItemId id, Function<BlockBehaviour.Properties, Block> block, BlockBehaviour.Properties properties) {
@@ -54,9 +60,9 @@ public interface RewiredBlocks {
 		properties.setId(blockKey);
 
 		//? fabric
-		//return Registry.registerForHolder(BuiltInRegistries.BLOCK, blockKey, block.apply(properties));
+		return Registry.registerForHolder(BuiltInRegistries.BLOCK, blockKey, block.apply(properties));
 		//? neoforge
-		return BLOCKS.registerBlock(blockKey.identifier().getPath(), block, () -> properties);
+		//return BLOCKS.registerBlock(blockKey.identifier().getPath(), block, () -> properties);
 	}
 
   static void init() {