+++ versions/fabric/src/client/java/dev/coder2195/rewired/datagen/RewiredDatagen.java
@@ -1,5 +1,6 @@
 package dev.coder2195.rewired.datagen;
 
+import dev.coder2195.rewired.datagen.loot_table.*;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 
@@ -8,5 +9,6 @@ public class RewiredDatagen implements DataGeneratorEntrypoint {
 	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
 		var pack = fabricDataGenerator.createPack();
 		pack.addProvider(ModelProvider::new);
+		pack.addProvider(BlockLootProvider::new);
 	}
 }