+++ CHANGELOG.md
@@ -10,4 +10,23 @@ This marks the initial alpha release of the project.
 **[+] OR Gate**
 - Releases level 15 when any enabled input has a signal, otherwise level 0
 
+**[+] XOR Gate**
+- Releases level 15 when an odd number of enabled inputs have a signal, otherwise level 0
 
+**[+] NAND Gate**
+- Releases level 15 when not all enabled inputs have a signal, otherwise level 0
+
+**[+] NOR Gate**
+- Releases level 15 when no enabled inputs have a signal, otherwise level 0
+
+**[+] XNOR Gate**
+- Releases level 15 when an even number of enabled inputs have a signal, otherwise level 0
+
+**[+] Average Gate**
+- Releases the average signal strength of all enabled inputs
+
+**[+] Min Gate**
+- Releases the minimum signal strength of all enabled inputs
+
+**[+] Max Gate**
+- Releases the maximum signal strength of all enabled inputs